REF: Move emulator into octue package
